Rating and Feedback system

+++ Assets/Scripts/Resources/Ingredient.cs
@@ -3,23 +3,22 @@
 [System.Serializable]
 public class Ingredient {
 
-
     /// <summary>
     /// Number of ingredients needed to cook, when it is used
     /// in the recepe.
     /// Or number of ingredients available in the storage.
     /// </summary>
-    public int Count = 0;
-    public int Max = 50;
+    public float Count = 0;
+    public float Max = 50;
 
-    public void Add(int amount) {
+    public void Add(float amount) {
         Count += amount;
         if (Count < 0)
             Count = 0;
     }//Add
 
 
-    public void Subtract(int amount) {
+    public void Subtract(float amount) {
         Add(-amount);
     }//Subtract
